modularize store & store
