+++ src/client/config.rs
@@ -1,0 +1,65 @@
+// CLIENT CONFIG
+// ================================================================================================
+
+use std::path::PathBuf;
+
+/// Configuration options of Miden client.
+#[derive(Debug, PartialEq, Eq)]
+pub struct ClientConfig {
+    /// Location of the client's data file.
+    pub store_path: String,
+    /// Address of the Miden node to connect to.
+    node_endpoint: Endpoint,
+}
+
+impl ClientConfig {
+    /// Returns a new instance of [ClientConfig] with the specified store path and node endpoint.
+    pub fn new(store_path: String, node_endpoint: Endpoint) -> Self {
+        Self {
+            store_path,
+            node_endpoint,
+        }
+    }
+}
+
+impl Default for ClientConfig {
+    fn default() -> Self {
+        const STORE_FILENAME: &str = "store.sqlite3";
+
+        // get directory of the currently executing binary, or fallback to the current directory
+        let exec_dir = match std::env::current_exe() {
+            Ok(mut path) => {
+                path.pop();
+                path
+            }
+            Err(_) => PathBuf::new(),
+        };
+
+        let store_path = exec_dir.join(STORE_FILENAME);
+
+        Self {
+            store_path: store_path
+                .into_os_string()
+                .into_string()
+                .expect("Creating the hardcoded path to the store file should not panic"),
+            node_endpoint: Endpoint::default(),
+        }
+    }
+}
+
+#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Debug)]
+pub struct Endpoint {
+    pub host: String,
+    pub port: u16,
+}
+
+impl Default for Endpoint {
+    fn default() -> Self {
+        const MIDEN_NODE_PORT: u16 = 57291;
+
+        Self {
+            host: "localhost".to_string(),
+            port: MIDEN_NODE_PORT,
+        }
+    }
+}
