+++ src/client/errors.rs
@@ -1,0 +1,28 @@
+use core::fmt;
+use objects::AccountError;
+
+use crate::store::errors::StoreError;
+
+#[derive(Debug)]
+pub enum ClientError {
+    StoreError(StoreError),
+    AccountError(AccountError),
+}
+
+impl fmt::Display for ClientError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            ClientError::StoreError(err) => write!(f, "store error: {err}"),
+            ClientError::AccountError(err) => write!(f, "account error: {err}"),
+        }
+    }
+}
+
+impl From<StoreError> for ClientError {
+    fn from(err: StoreError) -> Self {
+        Self::StoreError(err)
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for ClientError {}
