+++ src/main.rs
@@ -1,5 +1,4 @@
 use clap::Parser;
-use miden_client::{Client, ClientConfig};
 
 mod cli;
 use cli::Cli;
