+++ src/store/errors.rs
@@ -1,36 +1,6 @@
 use core::fmt;
 use crypto::utils::DeserializationError;
-use objects::{accounts::AccountId, AccountError, Digest};
-
-// CLIENT ERROR
-// ================================================================================================
-
-#[derive(Debug)]
-pub enum ClientError {
-    StoreError(StoreError),
-    AccountError(AccountError),
-}
-
-impl fmt::Display for ClientError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            ClientError::StoreError(err) => write!(f, "store error: {err}"),
-            ClientError::AccountError(err) => write!(f, "account error: {err}"),
-        }
-    }
-}
-
-impl From<StoreError> for ClientError {
-    fn from(err: StoreError) -> Self {
-        Self::StoreError(err)
-    }
-}
-
-#[cfg(feature = "std")]
-impl std::error::Error for ClientError {}
-
-// STORE ERROR
-// ================================================================================================
+use objects::{accounts::AccountId, Digest};
 
 #[derive(Debug)]
 pub enum StoreError {
